+++ src/TerraformPlugin.groovy
@@ -11,6 +11,7 @@
  */
 class TerraformPlugin implements TerraformValidateCommandPlugin,
                                  TerraformFormatCommandPlugin,
+                                 TerraformInitCommandPlugin,
                                  TerraformPlanCommandPlugin,
                                  TerraformApplyCommandPlugin,
                                  TerraformValidateStagePlugin,
@@ -25,6 +26,7 @@ class TerraformPlugin implements TerraformValidateCommandPlugin,
 
         TerraformValidateCommand.addPlugin(plugin)
         TerraformFormatCommand.addPlugin(plugin)
+        TerraformInitCommand.addPlugin(plugin)
         TerraformPlanCommand.addPlugin(plugin)
         TerraformApplyCommand.addPlugin(plugin)
         TerraformValidateStage.addPlugin(plugin)
@@ -75,6 +77,7 @@ class TerraformPlugin implements TerraformValidateCommandPlugin,
 
         TerraformValidateCommand.reset()
         TerraformFormatCommand.reset()
+        TerraformInitCommand.reset()
         TerraformPlanCommand.reset()
         TerraformApplyCommand.reset()
         TerraformValidateStage.reset()
@@ -87,6 +90,11 @@ class TerraformPlugin implements TerraformValidateCommandPlugin,
 
     @Override
     void apply(TerraformFormatCommand command) {
+        applyToCommand(command)
+    }
+
+    @Override
+    void apply(TerraformInitCommand command) {
         applyToCommand(command)
     }